Split providers from components
