+++ client/providers/ConnectionStatusProvider.js
@@ -1,8 +1,8 @@
 import { Meteor } from 'meteor/meteor';
 import React from 'react';
 
-import { ConnectionStatusContext } from '../../contexts/ConnectionStatusContext';
-import { useReactiveValue } from '../../hooks/useReactiveValue';
+import { ConnectionStatusContext } from '../contexts/ConnectionStatusContext';
+import { useReactiveValue } from '../hooks/useReactiveValue';
 
 export function ConnectionStatusProvider({ children }) {
 	const status = useReactiveValue(() => ({
